Revert "Merge branch 'main' into Ratatouillify"
This reverts commit 32cb6729e1229d9ccf5eae2d5dd3f19aec8cbebf, reversing
changes made to a368b2610e2200816edd7c7221a3ca4a17015037.
