+++ README.md
@@ -3,4 +3,3 @@
 Before running 'npm start' make sure to run:
 1. 'npm install --save react @trendmicro/react-sidenav' for nav bar package
 2. 'npm add react-router-dom" for ability to route to different pages from nav bar
-3. 'npm install --save react-helmet' for loading pages to redirect
