+++ ratatouille/src/App.css
@@ -223,110 +223,14 @@ img {
 .lessoncard {
   border: 5px solid var(--grey);
   background-color: white;
-  /* padding: 125px 330px; */
+  padding: 125px 330px;
   border-radius: 25px;
   margin: 7%;
-  position: absolute;
-  left: 16%;
-  display: table;
-  table-layout: fixed;
-  width: 55%;
-  height: 40%;
 }
 
 .lessoncard:hover {
-  box-shadow: var(--yellow) 0px 0px 30px;
-}
-
-.lessoncard-cell{
-  color: var(--grey);
-  display: table-cell;
-  width: 50%;
-  height: 100%;
-  align-self: start;
-  position: absolute;
-  left: 0;
-  top:0;
-}
-
-.lessoncard-image-cell{
-  display: table-cell;
-  width: 50%;
-  height: 100%;
-  position: absolute;
-  left: 50%;
-  top:0;
-}
-
-.lessoncard-image{
-  height: 80%;
-  width: 92%;
-  position: absolute;
-  top: 10%;
-  left: 0;
-  border-radius: 10px;
-}
-
-.headertext-lessoncard{
-  font-family: 'Open Sans';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 40px;
-  line-height: 44px;
-  color: var(--almost_black);
-  top: 10%;
-  position: absolute;
-  left: 10%;
-}
-
-.difficulty-label{
-  position: absolute;
-  left: 10%;
-  width: 20%;
-  top: 31.8%;
-  bottom: 55.08%;
-  background: rgba(44, 141, 53, 0.2);
-  border-radius: 50px;
-  font-family: 'Open Sans';
-  text-align: center;
-  font-style: normal;
-  font-family: 'Nunito Sans';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 24px;
-  line-height: 33px;
-  display: flex;
-  color: #2C8D35; 
-  text-align: center;
-}
-
-.difficulty-label-text{
-  position: absolute;
-  left: 15%;
-}
-
-.time-label{
-  position: absolute;
-  left: 35%;
-  top: 34%;
-  font-family: 'Nunito Sans';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 27px;
-  color: var(--grey);
-}
-
-.lessoncard-icons{
-  height: 40%;
-  width: auto;
-  position: absolute;
-  top: 50%;
-  left: 10%;
-}
-
-.row{
-  display: table-row;
+  box-shadow: var(--yellow) 0px 0px 20px;
+  padding: 127px 335px;
 }
 
 .yellow {
@@ -345,6 +249,7 @@ img {
   top: 20.7%;
   border-radius: 50%;
   position: absolute;
+  
 }
 
 .circle2 {
@@ -373,6 +278,7 @@ img {
   font-weight: 700;
   font-size: 32px;
   line-height: 44px;
+
   color: var(--almost_black);
 }
 
@@ -407,13 +313,13 @@ img {
 .home-newUser-text2{
   position: absolute;
   top: 25%;
-  left: 40%;
+  left: 48.5%;
 }
 
 .home-newUser-text3{
   position: absolute;
   top: 25%;
-  left: 51%;
+  left: 58%;
 }
 
 .orangetext {
@@ -440,130 +346,33 @@ img {
 
 .rathometext1{
   position: absolute;
-  left: 23%;
-  top: 100%;
+  left: 34%;
+  top: 95%;
 }
 
 .rathometext2{
   position: absolute;
-  top: 100%;
-  left: 43.5%;
+  top: 95%;
+  left: 54.5%;
 }
 
 .rathometext3{
   position: absolute;
-  left: 60%;
-  top: 100%;
+  left: 61.5%;
+  top: 95%;
 }
 
 .searchbar {
-  height: 70px;
-  box-sizing: border-box;
-  padding-left: 3.5rem;
-  align-items: center;
-  margin: 10px auto;
-  margin-top: 10px;
-  margin-right: auto;
-  margin-bottom: 10px;
-  margin-left: auto;
-  display: flex;
-  width: 55.6%;
+  position: absolute;
+  width: 53%;
+  left: 23%;
+  top: 82%;
+  height: 40px;
+  font-size: 24px;
   border-radius: 20px;
-  border: 10px solid var(--grey);
+  border: 5px solid var(--grey);
   background-color: white;
-  font-size: 24px;
-  color: var(--grey);
-  font-family: 'Nunito', sans-serif;
-  border-width: thick;
 }
-
-.home-search-wrapper {
-  left: 1.5%;
-  top: 80%;
-  align-items: center;
-  margin: 10px auto;
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 98%;
-
-}
-
-.home-search-icon {
-  left: 320px;
-  top: 0;
-  font-size: 25px;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--grey);
-  position: absolute;
-}
-
-.orangeButton{
-  background-color: var(--orange);
-  color: var(--almost_white);
-  font-weight: bold;
-  border: 3px solid var(--almost_black);
-  left: 43%;
-  border-radius: 20px;
-  position: absolute;
-  width: 14.6%;
-  height: 6.88%;
-  font-size: 24px;
-  position: absolute;
-  top: 108%;
-  margin-left: auto;
-  margin-right: auto;
-  font-family: 'Nunito Sans';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 24px;
-  line-height: 33px;
-}
-
-.rat-home-text{
-  text-align: center;
-  display: block;
-  position: absolute;
-  top: 115%;
-  left: 27.5%;
-}
-
-.small-text-home{
-  font-family: 'Nunito Sans';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 27px;
-  color: var(--almost_black);
-}
-
-.search-suggestions{
-  display: table;
-  width: 60%;
-  table-layout: fixed;
-  top: 93%;
-  left: 21%;
-  position: absolute;
-}
-
-.search-hint{
-  display: table-cell;
-  text-align: center;
-  font-family: 'Nunito Sans';
-  color: var(--orange)
-}
-
-.search-title{
-  display: table-cell;
-  text-align: center;
-  font-family: 'Nunito Sans';
-  color: var(--grey)
-}
-
 
 .searchbartop{
   position: absolute;
@@ -705,7 +514,7 @@ img {
 .link {
   color: white;
   font-size: 24px;
-  left: 45.5%;
+  left: 48.5%;
   position: absolute;
   top: 85%;
   font-family: 'Nunito', sans-serif;
@@ -893,72 +702,4 @@ margin:0
 }
 
 
-.tutorial-landing{
-  overflow-y: auto;
-  background-color: white;
-}
 
-.lesson-page-skills{
-  position: absolute;
-  width: 60%;
-  top: 20%;
-  left: 20%;
-}
-
-.tutorial-header{
-  position: absolute;
-  left: 20%;
-  top: 22.17%;
-  font-family: 'Open Sans';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 32px;
-  line-height: 44px;
-  display: flex;
-}
-
-.tutorial-equipments{
-  overflow-x: scroll;
-  top: 49%;
-  position: absolute;
-  background-color: aliceblue;
-  height: 180px;
-  left: 20%;
-  width: 60%;
-}
-
-.tutorial-ingredients{
-  overflow-x: scroll;
-  top: 79%;
-  position: absolute;
-  background-color: aliceblue;
-  height: 200px;
-  left: 20%;
-  width: 60%;
-}
-
-.tutorial-padding{
-  padding: 5% 5% 5% 5%;
-  position: absolute;
-  top: 100%;
-}
-
-.tutorial-start{
-  top: 10%;
-  right: 0;
-  position: fixed;
-  z-index: 900;
-}
-
-.tutorial-exit{
-  position: absolute;
-  top: 1%;
-  left: 90%;
-  height: 150%;
-  width: auto;
-  color: var(--grey)
-}
-
-.tutorial-exit:hover{
-  color: var(--orange);
-}
