+++ ratatouille/src/App.js
@@ -2,20 +2,18 @@ import logo from './logo.svg';
 import './App.css';
 import SideNavBar from './SideNavBar';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
-import Home_newUser from './pages/HomeNewUser';
+import Home_newUser from './pages/Home_newUser';
 import Lessons from './pages/Lessons';
 import Profile from './pages/Profile';
 import Ratatouillify from './pages/Ratatouillify';
 import Upload from './pages/Upload'
-import Login_landing from './pages/LoginLanding';
+import Login_landing from './pages/Login_landing';
 import Search from './pages/Search';
 import Login from './pages/Login';
 import Create_account from './pages/Create_account';
 import Home_returnUser from './pages/Home_returnUser';
 import Loading_recipe from './pages/Loading_recipe';
 import Lesson_Ready from './pages/Lesson_Ready';
-import Tutorial_landing from './pages/TutorialLanding';
-
 
 export default function App() {
   return (
@@ -34,7 +32,6 @@ export default function App() {
           <Route path="/upload" element={<Upload />}/>
           <Route path="/loading-recipe" element={<Loading_recipe />}/>
           <Route path="/lesson-ready" element={<Lesson_Ready />}/>
-          <Route path="/tutorial-landing" element={<Tutorial_landing />}/>
           <Route path="/*" element={<div>ERROR: Page Not Found</div>} />
         </Routes>
       </Router>
