+++ ratatouille/src/pages/Home_newUser.js
@@ -1,8 +1,6 @@
 import '../App.css';
 import {useNavigate} from 'react-router-dom';
 import Logo from '../Logo_orange.png';
-import lesson_icons from "../lessoncard-icons.png";
-import omelet from "../omelet.jpeg";
 
 function Home_newUser() {
     let navigate = useNavigate();
@@ -28,37 +26,16 @@ function Home_newUser() {
         <div class="circle3 yellow"></div>
         <span className='headertext hometext1' >Get started on a &nbsp;</span><span class='headertext home-newUser-text2 orangetext'>lesson plan</span><span class='headertext home-newUser-text3'>...</span>
         <br></br>
-        <button className='lessoncard' onClick={() => {navigate("/tutorial-landing")}}>
-          <div className='lessoncard-cell'>
-            <span className='row headertext-lessoncard'>Omelet</span>
-            <div className="difficulty-label"><div className="difficulty-label-text">easy</div></div>
-            <span className='time-label'>5 min</span>
-            <img src={lesson_icons} class='lessoncard-icons'></img>
-          </div>
-          <div className="lessoncard-image-cell">
-            <img src={omelet} className="lessoncard-image"></img>
-          </div>
-        </button>
+        <button class='lessoncard'>lesson card</button>
         <span className='headertext searchtext1' >Get help with something</span><span class='headertext searchtext2 orangetext'> &nbsp; specifc</span><span class='headertext searchtext3'>...</span>
         {/* search bar */}
-        <form onSubmit={()=> {navigate("/search");} } className="home-search-wrapper">
-          <span class="home-search-icon">
-            <i class="fa-solid fa-magnifying-glass"></i>
-          </span>
+        <form onSubmit={()=> {navigate("/search");}}>
           <input type='text' placeholder='Search...' class='searchbar'></input>
         </form>
-        <div className="search-suggestions">
-          <span className="small-text-home search-title">Top Suggested</span>
-          <span className="small-text-home search-hint">easy recipes</span>
-          <span className="small-text-home search-hint">knife skills</span>
-          <span className="small-text-home search-hint">pasta recipes</span>
-        </div>
         <span className='headertext rathometext1' >Upload a recipt to be </span><span className='headertext rathometext2 orangetext'> &nbsp; Ratatouilli-fied</span><span className='headertext rathometext3'>...</span>
         <button className="orangeButton" onClick={() => {navigate("/ratatouillify");}}>Ratatouillify</button>
-        <div className="rat-home-text">
-          <h3 className="small-text-home">Don’t see the recipe you’re looking for?</h3>
-          <h3 className='small-text-home'>We can Ratatouillify it and help you create an original lesson plan!</h3>
-        </div>
+        <span className="small-text-home">Don’t see the recipe you’re looking for?</span>
+        <span className='small-text-home'>We can Ratatouillify it and help you create an original lesson plan!</span>
       </div>
     </div>
   );
