+++ ratatouille/src/pages/Home_returnUser.js
@@ -1,0 +1,38 @@
+import '../App.css';
+import {useNavigate} from 'react-router-dom';
+import Logo from '../Logo_orange.png';
+
+export default function Home_returnUser() {
+    let navigate = useNavigate();
+    return (
+    <div class="row">
+      <div class="column left">
+        <img src={Logo} class='smallLogo'></img>
+        <div class="container">
+          <i class="fa fa-solid fa-user user" style={{ fontSize: '2em' }} />
+          <div class="btn-group">
+            <button onClick={() => {navigate("/login-landing")}}> Login </button>
+            <button onClick={() => {navigate("/")}}>Home</button>
+            <button onClick={() => {navigate("/profile");}}>Profile</button>
+            <button onClick={() => {navigate("/lessons");}}>Lessons</button>
+            <button onClick={() => {navigate("/ratatouillify");}}>Ratatouillify</button>
+          </div>
+        </div>
+      </div>
+      <div class="column right">
+        <img src={Logo} class='largeLogo'></img>
+        <div class="circle1 yellow"></div>
+        <div class="circle2 yellow"></div>
+        <div class="circle3 yellow"></div>
+        <span className='headertext hometext1 return' >Continue on your &nbsp;</span><span className='headertext hometext2 orangetext'>lesson plan</span><span class='headertext hometext3'>...</span>
+        <br></br>
+        <button class='lessoncard'>lesson card</button>
+        <span className='headertext searchtext1' >Get help with something</span><span class='headertext searchtext2 orangetext'> &nbsp; specifc</span><span class='headertext searchtext3'>...</span>
+        {/* search bar */}
+        <form onSubmit={()=> {navigate("/search");}}>
+          <input type='text' placeholder='Search...' class='searchbar'></input>
+        </form>
+      </div>
+    </div>
+  );
+}
